swap websocket transport class names
